THEMES/detorious: rework genlist theme.
Remove cruft (odd is now a signal, not a different group), refactor
the common part and simplify the theme.

I still don't like it much, rather go with a list like the one in
"everything" (e17), but until we settle on a look and feel, let's use
this.

Enjoy is now 50% better :-)



SVN revision: 62766